// src/types.ts
export enum IPCRendererActions {
  START_SCREENSHOTS = 'renderer/start-screenshots',
  END_SCREENSHOTS = 'renderer/end-screenshots',
  CHANGE_SCREENSHOT_INTERVAL = 'renderer/change-screenshot-interval'
}

export enum IPCMainActions {
  TOGGLE_RECORDING_STATE = 'main/toogle-state',
  SET_COUNTDOWN = 'main/set-countdown'
}